hardware: tcl script generation: move viado version setting around. Now, version can be preceised in te Config object, the vivado object looks it up and in case of auto, detect the version wanted and, finally, make sure a version is set when a method is called.
